requirements.txt
absl-py
mock
numpy
rouge-score
sacrebleu
sentencepiece
tensorflow-gpu==1.15
tensorflow-text==1.15.0rc0
tensor2tensor==1.15.0
tfds-nightly
